third_party/WebKit/LayoutTests/http/tests/security/postMessage/invalid-origin-throws-exception-expected.txt

CONSOLE ERROR: Unable to post message to asdf://. Recipient has origin http://localhost:8000.

window.location.href = http://127.0.0.1:8000/security/postMessage/invalid-origin-throws-exception.html
waiting...
Encountered exception "SyntaxError: Invalid target origin '' in a call to 'postMessage'." while posting message to ''.
Encountered exception "SyntaxError: Invalid target origin 'asdf' in a call to 'postMessage'." while posting message to 'asdf'.
Encountered exception "SyntaxError: Invalid target origin '/tmp/foo' in a call to 'postMessage'." while posting message to '/tmp/foo'.
Encountered exception "SyntaxError: Invalid target origin '//localhost' in a call to 'postMessage'." while posting message to '//localhost'.
Posted message to 'asdf:' without any exceptions.
Encountered exception "SyntaxError: Invalid target origin 'http:' in a call to 'postMessage'." while posting message to 'http:'.
Received message: data="Received message: data="done" origin="http://127.0.0.1:8000"" origin="http://localhost:8000"
